Fix doc slugger
